+++ src/main/java/com/spring/wsb/ecommerce/repositories/BasketItemRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface BasketItemRepository extends JpaRepository<BasketItem, Integer> {
+public interface BasketItemRepository extends JpaRepository<BasketItem, Long> {
 
-    Optional<BasketItem> findByProduct_Id(Integer id);
+    Optional<BasketItem> findByProduct_Id(Long id);
 }
